# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "dssd-unlp-2021-grupo11-paises-estados-service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/topEstados": {
            "get": {
                "tags": [
                    "estados"
                ],
                "summary": "Top Estados",
                "description": "Top Estados",
                "operationId": "getTopEstados",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/topIdiomas": {
            "get": {
                "tags": [
                    "idiomas"
                ],
                "summary": "Top idiomas",
                "description": "Top idiomas",
                "operationId": "getTopIdiomas",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/topContinente": {
            "get": {
                "tags": [
                    "continentes"
                ],
                "summary": "El continente del planeta (excluido América) hacia donde más se exporta",
                "description": "El continente del planeta (excluido América) hacia donde más se exporta",
                "operationId": "getTopContinente",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/continentesNoSeExporta": {
            "get": {
                "tags": [
                    "continentes"
                ],
                "summary": "Continentes hacia los que no se exporta",
                "description": "Continentes hacia los que no se exporta",
                "operationId": "getContinentesHaciaDondeNoSeExporta",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        },
        "/api/paisesNoSeExporta": {
            "get": {
                "tags": [
                    "paises"
                ],
                "summary": "Paises hacia los que no se exporta",
                "description": "Paises hacia los que no se exporta",
                "operationId": "getPaisesHaciaDondeNoSeExporta",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "500 internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "500 internal server error"
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}